backend/requirements.txt
django
psycopg2-binary
djangorestframework
markdown
django-filter
django-environ
djangorestframework-simplejwt
django-extensions
pillow
drf-extra-fields
weasyprint
xhtml2pdf
opencv-python 
Pillow 
numpy
corsheaders
